# src/main/resources/openapi/swagger-config.yaml
openapi: 3.0.1
info:
  title: Agroex API
  description: Spring Boot Application
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Lot controller
    description: Lot API
  - name: User controller
    description: User API
paths:
  /users/{id}:
    get:
      tags:
        - User controller
      summary: Get user by its id
      operationId: getById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully returned
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        "404":
          description: User with this id not found
    put:
      tags:
        - User controller
      summary: Update existing user by its id (must send all fields)
      operationId: updateUser_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: Successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        "404":
          description: User with this id not found
    delete:
      tags:
        - User controller
      summary: Delete user by its id
      operationId: deleteUserById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully deleted
        "404":
          description: User with this id not found
  /lots/{id}:
    get:
      tags:
        - Lot controller
      summary: Get lot by its id
      operationId: findById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully returned
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LotDTO'
        "404":
          description: Lot with this id not found
    put:
      tags:
        - Lot controller
      summary: Update existing lot (must send all fields)
      operationId: update_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotDTO'
        required: true
      responses:
        "404":
          description: Lot with this id not found
        "200":
          description: Successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LotDTO'
    delete:
      tags:
        - Lot controller
      summary: Delete lot by its id
      operationId: deleteById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully deleted
  /users:
    get:
      tags:
        - User controller
      summary: Get all created users
      operationId: getAll_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      tags:
        - User controller
      summary: Create user
      operationId: createUser_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: Successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
  /lots:
    get:
      tags:
        - Lot controller
      summary: Get all lots by its id
      operationId: findAll_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/LotDTO'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Lot'
    post:
      tags:
        - Lot controller
      summary: Create lot
      operationId: save_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotDTO'
        required: true
      responses:
        "200":
          description: Successfully created
          content:
            '*/*':
              schema:
               $ref: '#/components/schemas/LotDTO'
  /countries/{id}:
    get:
      tags:
        - Country controller
      summary: Get country by its id
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryDTO'
    put:
      tags:
        - Country controller
      summary: Update country by its id
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryDTO'
    delete:
      tags:
        - Country controller
      summary: Delete country by its id
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /categories/{id}:
    get:
      tags:
        - Product Category Controller
      summary: Get product category by its id
      operationId: findAllSubcategories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCategory'
    put:
      tags:
        - Product Category Controller
      summary: Update product category by its id (all fields must be sent)
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategory'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductCategory'
    delete:
      tags:
        - Product Category Controller
      summary: Delete product category by its id
      operationId: deleteById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /countries:
    get:
      tags:
        - Country controller
      summary: Get all countries
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryDTO'
    post:
      tags:
        - Country controller
      summary: Create country
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryDTO'
  /categories:
    get:
      tags:
        - Product Category Controller
      summary: Get category by its id
      operationId: findAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCategory'
    post:
      tags:
        - Product Category Controller
      summary: Create category
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategory'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductCategory'
components:
  schemas:
    UserDTO:
      required:
        - email
        - phoneNumber
        - username
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: 123_user_321
          maxLength: 64
          minLength: 3
        email:
          type: string
          example: user_test@gmail.com
        password:
          type: string
          writeOnly: true
          example: user-test123321
          maxLength: 2147483647
          minLength: 10
        phoneNumber:
          type: string
          example: "+375448901238"
        registrationDate:
          type: string
          format: date-time
          readOnly: true
          example: 2024-01-26T15:52:08.8822866Z
        emailVerified:
          type: boolean
          readOnly: true
          example: true
    Image:
      type: string
      example: 0776a0db-200e-40ca-a353-980b92785346.jpg
    ProductCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          example: Fruits

        parentId:
          type: integer
          format: int64
          example: 1
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: fresh
          maxLength: 64
          minLength: 1
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
        registrationDate:
          type: string
          format: date-time
        emailVerified:
          type: boolean
    LocationDTO:
      type: object
      properties:
        countryId:
          type: integer
          format: int64
          example: 1
        region:
          type: string
          example: "Minsk region"
          maxLength: 64
          minLength: 0
        latitude:
          type: string
          example: "53.902284"
        longitude:
          type: string
          example: "27.561831"
    LotDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Orange Lot"
          maxLength: 64
          minLength: 5
        description:
          type: string
          example: "Fresh oranges from the orchard"
          maxLength: 1000
          minLength: 0
        variety:
          type: string
          maxLength: 64
          minLength: 0
          example: "varient"
        size:
          type: string
          maxLength: 64
          minLength: 0
          example: "70+"
        packaging:
          maxLength: 64
          minLength: 0
          type: string
          example: "Container"
        enabledByAdmin:
          type: boolean
          example: true
        quantity:
          type: integer
          format: int32
          example: 10000
          minimum: 1
        pricePerTon:
          type: number
          format: float
          example: 1.5
        currency:
          maxLength: 10
          minLength: 0
          type: string
          example: "USD"
        creationDate:
          type: string
          format: date-time
          example: "2024-01-31T21:16:32.739066Z"
        expirationDate:
          type: string
          format: date-time
          example: "2024-01-31T21:16:32.739066Z"
        productCategoryId:
          type: integer
          format: int64
          example: 1
        lotType:
          pattern: \b(?:buy|sell)\b
          type: string
          example: "sell"
        userId:
          type: integer
          format: int64
          example: 1
        location:
          $ref: '#/components/schemas/LocationDTO'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
    CountryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        title:
          type: string
          example: "Belarus"
          maxLength: 64
          minLength: 1